% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMatrix.R
\name{corMatrix}
\alias{corMatrix}
\title{Calculate the correlation coefficient between two Matrices}
\usage{
corMatrix(X1, X2, equal = TRUE, method = "pearson", ncomp1, ncomp2, ...)
}
\arguments{
\item{X1}{matrix 1 (data.frame format is also accepted)}

\item{X2}{matrix 2 (data.frame format is also accepted)}

\item{equal}{whether the two matrices are consistent in direction, that is, the equivalent matrix.}

\item{method}{method of calculating correlation coefficient: "SMI", "RV", "RV2", "RVadj", "PSI", "pearson", "r1", "r2", "r3", "r4", "spearman", "GCD" or "kendall"}

\item{ncomp1}{maximum number of subspace components from the second matrix.}
}
\description{
To calculate the correlation coefficient between two equivalent matrices or non-equivalent matrices, refer to the specific method:

Indahl, U.G.; Næs, T.; Liland, K.H.; 2018. A similarity index for comparing coupled matrices. Journal of Chemometrics; e3049.
}
