% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMatrixTest.R
\name{corMatrixTest}
\alias{corMatrixTest}
\title{Hypothesis test on the correlation number between two matrices}
\usage{
corMatrixTest(
  X1,
  X2,
  alternative = "two.sided",
  method = "pearson",
  exact = NULL,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{X1}{matrix 1 (data.frame format is also accepted)}

\item{X2}{matrix 2 (data.frame format is also accepted)}

\item{alternative}{indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less". You can specify just the initial letter. "greater" corresponds to positive association, "less" to negative association.}

\item{method}{a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.}

\item{conf.level}{confidence level for the returned confidence interval. Currently only used for the Pearson product moment correlation coefficient if there are at least 4 complete pairs of observations.}

\item{ncomp1}{maximum number of subspace components from the first matrix.}

\item{ncomp2}{maximum number of subspace components from the second matrix.}
}
\description{
Hypothesis test on the correlation number between two matrices
}
