% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProsigA.R
\name{ProsigA}
\alias{ProsigA}
\title{Proteome Significance A}
\usage{
ProsigA(x, n)
}
\arguments{
\item{x}{A data.frame or matrix of proteomic data.}

\item{n}{Biological duplications number.}
}
\value{
P-value between each protein.
}
\description{
When there is no biological duplication (or only 2 biological duplications), calculate the p-value of the proteomic data.
}
\details{
Note: use empirical distribution.
}
