% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleWGCNA.R
\name{SimpleWGCNA}
\alias{SimpleWGCNA}
\title{A quick implementation function of WGCNA}
\usage{
SimpleWGCNA(
  dataExpr,
  PNcor = TRUE,
  useEmpiricalBeta = FALSE,
  FileName = "SimpleWGCNA result"
)
}
\arguments{
\item{dataExpr}{A data.frame of gene expression matrix}

\item{PNcor}{Logical value, whether the positive and negative correlations are considered separately}

\item{useEmpiricalBeta}{Logical value, whether to use empirical beta value}

\item{FileName}{The name of the output file}
}
\value{

}
\description{
The purpose of this function is to calculate the result of WGCNA simply and quickly based on the expression data. Most of the parameters refer to the original document of WGCNA, and a few parameters are set according to the author's own experience.
}
\details{
Note: This function cannot replace the complete WGCNA solution. If the purpose is not to get a quick overview but to draw a formal conclusion, please consider its applicability to WGCNA.
}
