% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentRnageDiffCI.R
\name{studentRangeDiffCI}
\alias{studentRangeDiffCI}
\title{Calculate simultaneous confidence intervals}
\usage{
studentRangeDiffCI(name, success, size, level = 0.95)
}
\arguments{
\item{name}{The name of each experiment}

\item{success}{The number of successful results in each experiment}

\item{size}{The number of times each experiment}

\item{level}{The level of Simultaneous confidence intervals}
}
\description{
To obtain simultaneous confidence intervals comparing several binomial parameters using the studentized range distribution with a score statistic.
}
\references{
Agresti A, Bini M, Bertaccini B, Ryu E. Simultaneous confidence intervals for comparing binomial parameters. Biometrics. 2008;64(4):1270-1275. doi:10.1111/j.1541-0420.2008.00990.x
}
